requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
Django==4.2.10
django-bootstrap-v5==1.0.11
django-extensions==3.2.3
djangorestframework==3.14.0
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.3.1
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
soupsieve==2.5
sqlparse==0.4.4
uritemplate==4.1.1
Werkzeug==3.0.1
